implement login flow for guest users
This commit implements the login flow for guest users in the PlaylistScreen.
When a guest user attempts to interact with a feature that requires authentication, they will be redirected to the login screen.
Upon successful login, the user's status will be updated, and they will be able to access the feature.
Error and success snackbars have also been added to provide feedback to the user during this process.