add PlayListWithItemsScreen
This commit introduces the `PlayListWithItemsScreen` composable.

This screen displays a list of playlists using a `LazyColumn`.
Each playlist is represented by a `MyListItem` composable, showing its title and media count.
The screen includes a `TopBar` with the title "My Lists".

A preview function, `PlayListWithItemsScreenPrev`, is also provided to visualize the screen with sample data in both light and dark themes.